ADT/Layoutlib: update API to control how layout expands.
Previous API was a single on/off to let the layouts expand at
render time depending on how much space they needed.

The new API can now control expansion is horizontal and/or
vertical (or not at all)

Basic implementation in the editor, with a manual detect of
"ScrollView" as top element. We should make the ViewRule handle
this somehow.

Change-Id: Idc503bc0d1d3df98fbf01cc84625952ca55a8afb