Fix nine patch size.
Since we moved the 9-patch chunk out of the NinePatch, the image
content portion is not in our control (it's owned by the default
implementation of NinePatchDrawable through an Android Bitmap),
and therefore should not include the control outer line.

This changeset makes the bitmap contains only the extracted
content and makes the computation of the patches take into account that
the draw method will be applied on a bitmap that is only
the content without the control line.
Basically the computation of the patches/fixed rectangle exclude the
outer control line.

Change-Id: I29c2224ee2311962b30d1fb155b85d5731d4e6ec