Add unit test configuration for moreUnit
Eclipse doesn't have a builtin mechanism to jump between a class and
its test, as well as a way to run tests associated with a given
class. There is however a plugin for this, "moreUnit".

Our projects have split off our unit tests in a separate project,
which requires some per-project configuration to tell moreUnit where
to look for tests. This information has to live with the project
rather than in my workspace, so I'm checking in the configuration file
- it is harmless if you don't have moreUnit but useful if you do.

Change-Id: Id29c46e2eaa89f2ad530c2b24700db28cc823b3b