ADT: Animation preparation.
Update layoutlib API to work better with the new scene
locking mechanism (for concurrent renderings), new error
types in SceneResult, and updated Animation listener.

ADT changes to record the view object in CanvasViewInfo,
and the current LayoutScene in ViewHierarchy.

Added a test menu item to start an animation. This
is temporary and will be fixed later.

Change-Id: I67df2d116afdfd23c093e4645d4a0f99695c5d95