ADT: Add support for some new resource types.
They are not technically new, but we never really
needed them. Turns out we do now (some are used by some widgets).

Also added an assert to detect new types of resources, and a test
to make sure that all resource types and folder types are added
to the folder-res relationship maps.

Change-Id: Ie4b6786ea6de9c69776794750732303b8d68b7bb