Change context menu label from "Clear" to "Default"
The property context menus in the layout editor lets you choose
between True and False (for booleans), and between the various
possible enum values for enum properties. You can also reset the
values back to whatever the defaults were.

The menu label for this was "Clear". This changeset changes this to
"Default" instead, since "Clear" sounds more like an action than a
persistent choice among many, and when this option is chosen you are
removing a local override of the property and picking up whatever the
default is.

Change-Id: Id3ea9618612ac44e2b965e77c9e84751364cf813