Fix threading issue with layout reload
The layout change listener (which can be called on any thread) must
not access SWT state unless it's on the SWT thread. There were some
existing SWT-redispatch blocks to handle this, but one code path was
missing. This changeset moves the entire method under a single SWT
redispatch block instead.

Change-Id: I7d802d78fbdd4811c68de830e1a54b1a97d76b5f