Clean up the api around Layoutlib.
Move (Style/DensityBased)ResourceValue into layoutlib_api
and make the API use that instead of the interface.
We'll get ride of the interfaces once only obsolete platforms
use them.

In ide-commons also got rid of LayoutBridgeWrapper and moved
the code in LayoutLibrary which does not expose the bridge
anymore, and instead expose an API similar to the LayoutBridge
class.

Updated ADT to use LayoutLibrary directly instead of going through
LayoutLibrary.getBridge(). This allows us to hide some
things like querying the API level and relying instead on
Capabilities (with special handle for legacy bridges).

Also added an error message to LayoutLibrary to display why
it may have failed to load.
Added a check to the API level and don't load layoutlib
that are more recent than the client.

Change-Id: Ie4e615d8d32485ee577bb88e95cd3f562bf590cb