Various layout fixes
First, fix absolute layout such that it properly handles both screen
scaling as well as various screen resolution densities when it
computes the dip positions.

Second, pass the bounds surrounding the mouse position when a drag is
initiated to the view rules. This is used by both absolute layout and
relative layout to properly handle dragging bounds. In particular, in
relative layout this is used to match a border not only when the mouse
cursor gets near the edge, but when the bound edges also get near the
edge. In absolute layout this is used to show a correct bound
rectangle that has the proper offset from the original drag position
(which may not be the center, which until now it was assuming).

Third, in RelativeLayout, when there are no children, offer a left
alignment regardless of where you are within the rectangle. This is
similar to how LinearLayout works.

In addition, two internal changes:
* Factor the various RelativeLayout string constants into the
  LayoutConstants class. This had the nice side effect of revealing a
  typo where we were referencing a non-existing value!  Constants FTW!
* Make the RulesEngine log the exceptions, not just the error
  messages, thrown by IViewRule calls. That way the full stack trace
  is available in the Error Viewer, including line numbers etc.

Change-Id: I0b83df71b36741e65a1eb2003ed044157eb6f0cd