Set layout editor scrollbar increments
Fix 13097: Gingerbread Layout Editor scroll bar bugs, part 1:
Set the layout editor scrollbar increments to 20 (they were 1, meaning
that each click on the scrollbar down button would scroll the view by
one pixel rather than a "chunk" like 20).

Set the scrollbar page increment to the visual height of the
canvas. This means that if you click below the visible range in the
scrollbar it will jump a full scrollbar increment, as expected.

(This is the first half of 13097. The second half deals with something
unrelated.)

Change-Id: I80c544b749d08d1fdc127d682ef8a75aed9aa8e7