Fix DDMS and HierarchyViewer .bat files on Windows
These were not setting the "bin dir" properpty correctly,
and as a consequence the tools complain they can't find adb.exe.

The fix is to set it to the directory of the script but to
avoid spaces-in-path issues we use the "~fs" modifier to get
a space-free short name for the %cd% value. And since we can't
use use %cd% combined with ~fs, we use a for loop trick to set it.

Note that ddms.bat was in unix end-of-line mode so I changed it
back to DOS end-of-line mode.

Change-Id: Ic9e55643b551898043add27c12193a0949f40bb0