Fix Palette scrollbar bug
Fixes the problem that on Mac, when you close a palette category the
scrollbar thumb size was not updated. The thumb value could be set to
0 in that case, which is ignored by SWT.

This changeset also sets the increment and the page increment on the
scrollbar to improve the behavior of clicking the scrollbar arrows or
clicking above and below the thumb.

The changeset also sets all the values of the scrollbar in a single go
(via the setValues() call); this should improve performance, and in
conjunction with the first fix (to the thumb value) I am hoping this
addresses issue 13097: Gingerbread Layout Editor scroll bar bugs

Change-Id: I8274e4bdd8828ac814f04b99650f1209b8636797