Fix keyboard shortcut bug
13231: Keyboard shortcuts often don't work in the layout editor

We've had this bug where intermittently the keyboard shortcuts for
cut, copy, paste, undo, etc. haven't worked.  Turns out it's not
intermittent at all: Our keybinding registration was wrong, which
meant that only the last opened editor would actually work; the global
actions would ALWAYS map to that one editor. I discovered this when I
happened to have two editors side by side!

The global action registration mechanism we were using assumes that
each view has its own IActionBar. That's generally true, but NOT for
editors! That means we have to re-register the action bindings
ourselves when the active editor changes. To do that, we register our
own IEditorActionBarContributor, which will notify us when the actions
need to be updated for editor activation.

Change-Id: Ia49246a1f0fd14a717efe591b04b1b6529b23922