New "DecorComposite" for sub-panels in the GLE.
A DecorComposite wraps a control or composite such as the palette
and adds a little header with an icon, a title and an optional
toolbar of icon/actions.

The GLE palette is revamped to use this. Some actions will be
added later to toggle viewing views or layouts.

Eventually the new property panel will be one of these too.

Terminology: I tried to avoid the words "panel" or "page" or "pane"
which are just too overused. Since this is merely more than decoration
around a control, it's a DecorComposite.

I've also added a couple builder classes to make it less verbose
to set GridLayout and GridData objects. If these are useful out
of the editors, we could move them somewhere else later.

Change-Id: I5065411bfe48148c267ec6438b0ec5424123ab8c