SDK Manager in ADT: make it possible to unload the SDK.
One of the issues when using the SDK Manager from Eclipse
is that, at least on Windows, we can't upgrade the platform-tools
or the targets since they might be locked by ADT.
Typically dex.jar is in use and typically we have various data/res
(ttf and xml) files used by any opened layout editor.

This adds the necessary infrastructure to know in ADT when
packages are going to be installed. There's a crude attempt to
solve the dex.jar issue. However unloading the targets isn't
done yet.

Change-Id: I5c48144501c7f39ef779bfeffdfae85a48c65a29