Fix NON-NLS tokens
There were a number of // $NON-NLS-1$ references in the codebase.
Eclipse's "Externalize Strings" functionality will not handle these
correctly; there must not be a space between the "//" and the "$NON"
tokens.

(I left AndroidXmlEditor.xml alone; it is the file I discovered
the problem in but I fixed those references as part of another
pending checkin.)

Change-Id: If185c88a667273af614f0bee5959fd2618756c05