Improve the Outline contents
1. Make the outline use StyledStrings such that we can use different
colors for different elements in the outline. Use the decorations
color for the element type that follows the id.

2. For elements that define a "text" property, include the text (or a
prefix of it if it is long) in the outline. Thus, for a Button you
might see something like "Button01 - "Submit Order").

3. For elements that define a "src" property, show the
source. Therefore, for an ImageView you might see "ImageView - logo".

4. For <include> elements, show the name of the included layout.

Change-Id: Ibd4c8339ea0e03c969ccaec1a67bc64436ed67af