Replace Sdk.makeRelativeTo()
IPath#makeRelativeTo(IPath) was not supported on Eclipse 3.4, so we
had a local version of it in our sourcebase, as
Sdk#makeRelativeTo(IPath,IPath). However, our version only works
correctly for directories, not plain files, but I had been using it
for files as well.

Now that we no longer need to support Eclipse 3.4, remove our local
version and use the builtin path conversion method.

(Fixed some invalid javadoc too)

Change-Id: I233875e1ecc758eb1ed333686b319b138eb47c4a