Add "Play Animation" menu
Adds a new dynamically populated context menu named "Play Animation", which
contains 4 sections:

- First, the most recently played animation name. This is initially
  empty, but once you've played an animation, the most recently one is
  listed there.

- "Create...". This is a shortcut which opens the "New Android XML
  File" wizard, pre-configured to point at the current project and
  with the Animation category selected.

- "Android Builtin >". This is a submenu (also lazily populated) which
  lists all the builtin Android framework animations alphabetically.

- The animations in the current project, if any, listed
  alphabetically.

Selecting any of the project or framework animation names will cause
the animation to be played once. During the animation, selection and
hover overlays are temporarily hidden.

This is just a start; we should play/rewind/stop buttons, maybe a loop
toggle button, etc somewhere on the screen when you've selected an
aniation. We should also try to filter out animations (especially
framework ones) that have no effect on the current selection.
Finally, error messages need to be handled better since we are reusing
the render session from the editor.

Change-Id: If113b02d458972b5cc2cb253dd9e6c312b8302a9