Remove error(tag, throwable) from LayoutLog.
We should never only log an exception with no associated
message.
Also, I've found several case in the layoutlib where there
is a message in place of the tag.

This new API makes it more clear that there's both a tag and a message
since both are required for error()

Change-Id: I6d5f45c07b1cb8df96311d930170526bb729fae6