Choose a good locale when opening a file for the first time.
Because the selection of the full config to use for display
is complicated (due to having to find something that matches
the configuration of the edited file), this is a bit more
complex than simply changing the select in updateLocales().

In fact the selection made there when there is no stored config
is pointless. Instead it's decided later by findAndSetCompatibleConfig
which find the proper combination of device and locale*

* this is actually broken since we added the nightmode and dock mode.
We need to include those in the list of config we try out. I'll fix
this in a later patch.

Change-Id: I83599dc7f0d1a921f9568b6de5c065ca1e82e103