Remove <merge> from the palette
The <include> tag is already removed from the palette. This changeset
does the same thing for the similar <merge> tag, since it does not
work; you cannot drag a <merge> tag somewhere into an existing layout.
(The common case of rendering views with <merge> as the root -does-
work, but that does not involve the palette.)

Change-Id: I9210e196ab66fe77b1303e756384844fde1d1c0f