Improve hyperlink resolution for configurations
The hyperlink resolver would only look for values in the base values/
folder. This did not work well for resources that are ONLY defined in
specific configurations. This changeset improves the search to look in
all eligible resource folders. It also uses support from the layout
library to more quickly identify the right files to check.

The hyperlink resolver now also considers the configuration chooser in
the designtab (if opened), and uses this configuration
first. Therefore, if you for example have chosen to view a particular
language, hyperlink resolution will jump to the specific translation
string rather than the base string as before.

A few other fixes are included:

- Look in the current file for @id references to resolve first (common
  in a layout where you have @id references to attachments)
- Fix search for @attr/ attributes
- Fix bug where attributes inside a custom view class would be ignored
  and it would jump to the custom class instead

Change-Id: Ia97a3f45bf454cc28378387b219af0fdd855a902