Combine missing class errors with logger errors
When there are missing custom classes, or classes that cannot be
instantiated, a special error display is shown with hyperlinks to the
classes. However, this view does not incorporate the other logging
errors, such as resource failures.

This changeset combines the output so that you see everything --
missing and broken classes, resource warnings, and any other rendering
problems.

It also makes the hyperlinks pointing to classes in the error output
*open* the class if it already exists, and if not, continue to open
the New Class wizard as before.

Change-Id: Ic79282b8f502f03f587028040150e2cca3395fd9