Reset the ProjectCallback when the rendering target changes.
This is only done if the layout has custom views. This is
because ProjectCallback caches classes, but they are loaded
using the layoutlib classloader for type provided by the
layoutlib.

Since the rendering target changed, the layoutlib changed,
and we need a version of the custom view classes that uses
the base types of the new layoutlib jar.

Change-Id: Ia3e2a8f65d73dab2769d8c7686f970e647432ed3
http://code.google.com/p/android/issues/detail?id=14061