Add palette category metadata
This changeset contains some background work for the palette preview
work, separated out to make that changeset smaller.

First, it creates a new metadata XML file, which augments the metadata
provided by the platform, defining things like palette categories, a
natural order for the views within each category. The older code-based
metadata for fill preferences (used to decide how to set the width and
height attributes on drop) are also moved into this XML file. There
will be more metadata added to this file in the next changeset, where
for example XML fragments defining how to render a view for preview
purposes will be defined there.

Second, change the signature on the view and layout descriptor lists
passed around such that we don't have to do instanceof
ViewElementDescriptor in various places.

Fix sentence capitalization for a couple of undo label strings.

Change-Id: I66191ccf4cc0f4105c2331d496f3674ac1ad8b9d